user requesr refactor
