+++ Nol3.Communication.Models/Nol3.Communication.Models.csproj
@@ -42,10 +42,13 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="NOL3RegistrySetting.cs" />
-    <Compile Include="NolAPI\RequestType.cs" />
-    <Compile Include="NolAPI\ROOTFIXML.cs" />
-    <Compile Include="NolAPI\UserRequest.cs" />
+    <Compile Include="NolAPI\Requests\RequestType.cs" />
+    <Compile Include="NolAPI\Requests\ROOTFIXML.cs" />
+    <Compile Include="NolAPI\Requests\UserRequest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="NolAPI\Responses\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
