+++ Nol3.Communication/FIXML.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Nol3.Communication.Models;
-using Nol3.Communication.Models.NolAPI;
+using Nol3.Communication.Models.NolAPI.Requests;
 using System.IO;
 using System.Xml;
 using System.Xml.Serialization;
@@ -34,22 +34,9 @@ namespace Nol3.Communication
 
 			return result;
 		}
-		//public static string GenerateUserRequest(int requestType, int requestId)
-		//{
-		//	ROOTFIXML<TestClass> request = new ROOTFIXML<TestClass> { };
-		//	StringWriter stringWriter = new StringWriter();
-		//	XmlWriter xmlWriter = XmlWriter.Create(stringWriter, new XmlWriterSettings { OmitXmlDeclaration = true });
-
-		//	XmlSerializer ser = new XmlSerializer(typeof(ROOTFIXML<TestClass>));//, overrides);
-		//	XmlSerializerNamespaces xmlns = new XmlSerializerNamespaces();
-		//	xmlns.Add("", "");
-
-		//	using (xmlWriter)
-		//	{
-		//		ser.Serialize(xmlWriter, request, xmlns);
-		//	}
-
-		//	return stringWriter.ToString();
-		//}
+		public static string GenerateUserRequest(UserRequest userRequest)
+		{
+			return GenerateRequest<UserRequest>(userRequest);
+		}
 	}
 }
